SMA_CharacterizationV2/requirements.txt
pyvisa>=1.13
pyserial>=3.5
pyyaml>=6.0
numpy>=1.24
matplotlib>=3.7
readchar>=4.0
